+++ CHANGELOG.md
@@ -4,7 +4,9 @@
  - You can now configure the graves to only be loot-able when holding a shovel. Default disabled
  - You can configure probability of any entity spawning when looting a grave. Default disabled (0%)
    - The entity can be configured to which entity, and also completely custom NBT
-   - Nbt can contain special values representing player name (of grave owner), player UUID, and/or any item in grave
+   - Nbt can contain special values representing player name (of grave owner or looter), player UUID
+   - Nbt can also contain any item (based on which slot) in the grave, and even if this item should be "used up" (deleted from grave) or not
+ - Improved grave area overrides to handle keep inventory zones
 
 ### Fixes
  - Grave scroll is now craft-able when not disabled